backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic-ai==0.0.10
python-dotenv==1.0.0
httpx==0.25.2
cors==1.0.1
pydantic==2.5.0
openrouter==0.1.0
